requirements.txt
colorama==0.4.6
PyYAML==6.0